Move Todos from `App C` to `Todos C`
When I create components, I tend to hardcode the data and then stick
it in the root component for easy viewing and not needing a route
to look at it. The downside of this is that sometimes I forget to
place it in its respective component.

Addresses #6 (GET Todos)
Project: 'Todos CRUD'