Delete todo by id with click handler
IMO making individual todo item components that are iterated over
in a todos component like I have now is overkill, so rather than
the binding that the turorial I am skeletoning my stuff on using that
I just use a click handler for now. I get that granularity in React,
but with the seperation of HTML from JS in Angular that feels like
overkill. If the individual logic of `Todos C` or its HTML gets
bloated later it may be worth it.

Side note: JSONPlaceholder that I am using for the API doesn't actually
delete the resource, but it mocks the REST effects of doing so. Refreshing
will GET all elements again, so the element will be rerendered.

Addresses #3 (DELETE Todos)
Project: 'Todos CRUD'